Move the "A/B" testing options to end
When we add support for "A/B" testing of different message-bodies, there
will be more ways to enable "A/B" testing.  We may not know on the first
screen whether "A/B" is active.

This moves the "A/B" distribution options to the end -- at which point, we
will know for sure whether "A/B" testing is active.